feat: use remark-react to parse markdown to ReactNode
